Create modified_drupal_http_request.php
This is a modified version of the drupal_http_request function provided by the includes/common.inc file
This version contains added functionality to handle redirect headers that contain relative paths. 
So where a non relative redirect path like  http://somedomain.com/some_file.php has no problems
a redirect path that was simply /some_file.php would cause a failure in the original drupal_http_request
In this version - I added an assumption that if the redirect location doesnt parse a uri scheme (which would fail)
to use the original uri scheme and host values to build a fully qualified url to the relative path. 
this works for me because I was making requests that urls that redirect to relative paths - so while my entry point was working
drupal kept throwing 'missing schema' errors in watchdog as the redirects would fail.